# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A book management service API in Go using Gin framework.",
        "title": "Gin Book Service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/address/add": {
            "post": {
                "description": "Add address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Address"
                ],
                "summary": "Add address",
                "parameters": [
                    {
                        "description": "address",
                        "name": "address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/form.AdressForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "add address ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/address/all": {
            "get": {
                "description": "Get all address",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Address"
                ],
                "summary": "return list of all",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Address"
                        }
                    }
                }
            }
        },
        "/address/delete/{id}": {
            "delete": {
                "description": "Delete address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Address"
                ],
                "summary": "Delete address",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "DELETE ADDRESS OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/company/add": {
            "post": {
                "description": "Add company",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Add company",
                "parameters": [
                    {
                        "description": "company",
                        "name": "company",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/form.Company"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "add company ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/company/all": {
            "get": {
                "description": "Get all company",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "return list of all",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Company"
                        }
                    }
                }
            }
        },
        "/company/delete/{id}": {
            "delete": {
                "description": "Delete company",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Delete company",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "DELETE OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "login",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/form.FormLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/staff/add": {
            "post": {
                "description": "Add staff",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Staff"
                ],
                "summary": "Add staff",
                "parameters": [
                    {
                        "type": "file",
                        "description": "staff",
                        "name": "staff",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "image",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "add staff ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/staff/all": {
            "get": {
                "description": "Get all staff",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Staff"
                ],
                "summary": "return list of all",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Staff"
                        }
                    }
                }
            }
        },
        "/staff/delete/{id}": {
            "delete": {
                "description": "Delete staff",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Staff"
                ],
                "summary": "Delete staff",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "DELETE OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/staff/{id}": {
            "get": {
                "description": "Get all staff by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Staff"
                ],
                "summary": "return list of all",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Staff"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "form.AdressForm": {
            "type": "object",
            "properties": {
                "address_company": {
                    "type": "string"
                },
                "address_specific": {
                    "type": "string"
                },
                "company_id": {
                    "type": "integer"
                }
            }
        },
        "form.Company": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "form.FormLogin": {
            "type": "object",
            "properties": {
                "pass_word": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "models.Address": {
            "type": "object",
            "properties": {
                "address_company": {
                    "type": "string"
                },
                "address_specific": {
                    "type": "string"
                },
                "company": {
                    "$ref": "#/definitions/models.Company"
                },
                "company_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "staff": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Staff"
                    }
                }
            }
        },
        "models.Company": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Address"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "staff": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Staff"
                    }
                }
            }
        },
        "models.QR": {
            "type": "object",
            "properties": {
                "filePatch": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "staff": {
                    "$ref": "#/definitions/models.Staff"
                },
                "staffId": {
                    "type": "integer"
                }
            }
        },
        "models.Staff": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/models.Address"
                },
                "address_id": {
                    "type": "integer"
                },
                "company": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Company"
                    }
                },
                "email": {
                    "type": "string"
                },
                "fax": {
                    "type": "string"
                },
                "file_path_avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mobile_phone": {
                    "type": "string"
                },
                "name_company": {
                    "type": "string"
                },
                "qr": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.QR"
                    }
                },
                "role": {
                    "type": "string"
                },
                "staff_name": {
                    "type": "string"
                },
                "tele_phone": {
                    "type": "string"
                }
            }
        }
    }
}